+++ api/V0GetKubeVersion.go
@@ -26,21 +26,21 @@ func V0GetKubeVersion(w http.ResponseWriter, r *http.Request) {
 	res, err := client.Do(req)
 
 	if err != nil {
-		logrus.Error("GetKubeVersion: Error 1: ", err, res)
+		logrus.Error("V0GetKubeVersion: Error 1: ", err, res)
 		return
 	}
 
 	defer res.Body.Close()
 
 	if res.StatusCode != 200 {
-		logrus.Error("GetKubeVersion: Error Status is not 200")
+		logrus.Error("V0GetKubeVersion: Error Status is not 200")
 		return
 	}
 
 	content, err := ioutil.ReadAll(res.Body)
 
 	if err != nil {
-		logrus.Error("GetKubeVersion: Error 2: ", err, res)
+		logrus.Error("V0GetKubeVersion: Error 2: ", err, res)
 		return
 	}